pyproject.toml
[project]
name = "dl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pyautogui>=0.9.54",
    "pynput>=1.8.1",
    "scikit-image>=0.25.2",
]
